requirements.txt
fastapi
uvicorn
numpy
pandas
scikit-learn
joblib
python-dotenv
google-generativeai
